FIX: Provide non-zero exit code when an action fails
We were catching the exception and printing the message to format
the string nicer and not raise errors from the cli. But, when testing
with another command line program we weren't getting a known error due
to the zero exit code. This moves the handling of the try/except up
into the parser and just directly exits with a non-zero exit code.